indexer/requirements.txt
qdrant-client
sentence-transformers
pdfplumber
langchain
langchain-text-splitters
python-dotenv